add edge route, update exportProject to edge runtime, add repo directory and routes for all repo funcs
