+++ src/routes/login/+server.js
@@ -1,3 +1,4 @@
+// eslint-disable-next-line import/no-unresolved
 import { redirect } from '@sveltejs/kit';
 import { v4 as uuidv4 } from 'uuid';
 
@@ -19,7 +20,7 @@ export async function GET({ url, cookies }) {
   const repoName = searchParams.get('repoName');
 
   // Store the application URI for redirect URL
-  const redirectURI = 'http://localhost:5173/callback';
+  const redirectURI = 'https://svetch.vercel.app/callback';
 
   // Define the OAuth scope to be requested
   const OAuthScope = 'scope=repo%20read:user%20user:email';
@@ -32,8 +33,6 @@ export async function GET({ url, cookies }) {
 
   // Construct complete redirectURL
   const redirectURL = `${target}?client_id=${clientID}&redirect_uri=${redirectURI}/${redirectParams}`;
-
-  console.log('state:', state)
 
   cookies.set('state', state, {path: '/', HttpOnly: true})
