+++ src/routes/repo/createForAuthenticatedUser/+server.js
@@ -1,0 +1,18 @@
+import { Octokit } from 'octokit';
+import { restEndpointMethods } from '@octokit/plugin-rest-endpoint-methods';
+
+// eslint-disable-next-line import/prefer-default-export
+export async function POST({ request }) {
+  
+  // Parse body of incoming request object
+  const { repo, token } = await request.json();
+
+  // Store a custom Octokit object with rest API plugin
+  const CustomOctokit = Octokit.plugin(restEndpointMethods);
+
+  // Store an initialized Octokit (pseudoclassical) object
+  const octokit = new CustomOctokit({ auth: token });
+
+  await octokit.rest.repos.createForAuthenticatedUser({ name: repo, auto_init: true });
+  return new Response(200)
+}
